backend/requirements.txt
fastapi
mysql-connector-python
python-dotenv
uvicorn
bcrypt
pyjwt
passlib[bcrypt]